+++ packages/data-schema/src/SchemaProcessor.ts
@@ -76,7 +76,8 @@ function isInternalModel(model: unknown): model is InternalModel {
   if (
     (model as any).data &&
     !isCustomType(model) &&
-    !isCustomOperation(model)
+    !isCustomOperation(model) &&
+    !isConversationRoute(model)
   ) {
     return true;
   }
@@ -1497,7 +1498,6 @@ const schemaPreprocessor = (
             break;
         }
       } else if (isConversationRoute(typeDef)) {
-        // TODO: add inferenceConfiguration values to directive.
         const { field, functionHandler } = createConversationField(
           typeDef,
           typeName,