+++ packages/integration-tests/CHANGELOG.md
@@ -1,5 +1,17 @@
 # integration-tests
 
+## 0.2.0
+
+### Minor Changes
+
+- 068f8af: Fix: change accumulator return type for initializeModel function
+
+## 0.1.3
+
+### Patch Changes
+
+- 7d3c2bf: fix: Export ResourceAuthorization to fix related declaration errors
+
 ## 0.1.2
 
 ### Patch Changes